more refactoring to SearchListView
